+++ relay/client_test.go
@@ -86,9 +86,6 @@ func TestRelayClientEndToEnd(t *testing.T) {
 	clientOpts := smb.Options{
 		Host:              relayAddr.IP.String(),
 		Port:              relayAddr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,