+++ relay/fake_server_test.go
@@ -83,9 +83,6 @@ func TestFakeServerRoundTrip(t *testing.T) {
 	opts := smb.Options{
 		Host:              relayAddr.IP.String(),
 		Port:              relayAddr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,
@@ -208,9 +205,6 @@ func TestFakeServerReadOnly(t *testing.T) {
 	opts := smb.Options{
 		Host:              relayAddr.IP.String(),
 		Port:              relayAddr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,