+++ relay/forward_test.go
@@ -87,9 +87,6 @@ func TestSMBForwarderRawNTLMBoundary(t *testing.T) {
 	bait := smb.Options{
 		Host:              relayAddr.IP.String(),
 		Port:              relayAddr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,
@@ -172,9 +169,6 @@ func TestSMBForwarderMICStripDoesNotBreakRelay(t *testing.T) {
 	bait := smb.Options{
 		Host:              relayAddr.IP.String(),
 		Port:              relayAddr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,