+++ smb/server/preauth_test.go
@@ -109,9 +109,6 @@ func TestPreauthAntiDowngrade(t *testing.T) {
 		opts := smb.Options{
 			Host:                  "127.0.0.1",
 			Port:                  addr.Port,
-			User:                  user,
-			Password:              password,
-			Domain:                domain,
 			Initiator:             &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 			RequireMessageSigning: true,
 			DisableEncryption:     true,