+++ smb/server/sessionsetup_test.go
@@ -55,9 +55,6 @@ func TestSessionSetupCredentialCapture(t *testing.T) {
 	opts := smb.Options{
 		Host:              "127.0.0.1",
 		Port:              addr.Port,
-		User:              "bogus",
-		Password:          "nopass",
-		Domain:            "BOGUSDOM",
 		Initiator:         ntlmInit,
 		DisableSigning:    true,
 		DisableEncryption: true,
@@ -176,9 +173,6 @@ func TestSessionSetupMapAuthSuccess(t *testing.T) {
 	opts := smb.Options{
 		Host:              "127.0.0.1",
 		Port:              addr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,