+++ smb/server/signing_test.go
@@ -56,9 +56,6 @@ func TestSigning311(t *testing.T) {
 	opts := smb.Options{
 		Host:                  "127.0.0.1",
 		Port:                  addr.Port,
-		User:                  user,
-		Password:              password,
-		Domain:                domain,
 		Initiator:             &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		RequireMessageSigning: true,
 		DisableEncryption:     true,
@@ -126,9 +123,6 @@ func TestSigning21(t *testing.T) {
 	opts := smb.Options{
 		Host:                  "127.0.0.1",
 		Port:                  addr.Port,
-		User:                  user,
-		Password:              password,
-		Domain:                domain,
 		Initiator:             &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		RequireMessageSigning: true,
 		DisableEncryption:     true,
@@ -206,9 +200,6 @@ func TestUnsignedRejected(t *testing.T) {
 	opts := smb.Options{
 		Host:                  "127.0.0.1",
 		Port:                  addr.Port,
-		User:                  user,
-		Password:              password,
-		Domain:                domain,
 		Initiator:             &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		RequireMessageSigning: true,
 		DisableEncryption:     true,