+++ smb/server/treeconnect_test.go
@@ -99,9 +99,6 @@ func TestTreeConnect(t *testing.T) {
 	opts := smb.Options{
 		Host:              "127.0.0.1",
 		Port:              addr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,
@@ -151,9 +148,6 @@ func TestTreeConnectBadShare(t *testing.T) {
 	opts := smb.Options{
 		Host:              "127.0.0.1",
 		Port:              addr.Port,
-		User:              user,
-		Password:          password,
-		Domain:            domain,
 		Initiator:         &spnego.NTLMInitiator{User: user, Password: password, Domain: domain},
 		DisableSigning:    true,
 		DisableEncryption: true,